refactor(headless): isolate product-listing bundle
https://coveord.atlassian.net/browse/KIT-2461
