+++ packages/headless/src/controllers/core/facets/range-facet/numeric-facet/headless-core-numeric-filter.ts
@@ -1,16 +1,12 @@
 import {configuration} from '../../../../../app/common-reducers';
 import {CoreEngine} from '../../../../../app/engine';
 import {
-  facetOptions,
-  numericFacetSet,
-  search,
-} from '../../../../../app/reducers';
-import {
   disableFacet,
   enableFacet,
   updateFacetOptions,
 } from '../../../../../features/facet-options/facet-options-actions';
 import {isFacetEnabledSelector} from '../../../../../features/facet-options/facet-options-selectors';
+import {facetOptionsReducer as facetOptions} from '../../../../../features/facet-options/facet-options-slice';
 import {isFacetLoadingResponseSelector} from '../../../../../features/facets/facet-set/facet-set-selectors';
 import {NumericFacetValue} from '../../../../../features/facets/range-facets/numeric-facet-set/interfaces/response';
 import {
@@ -19,6 +15,8 @@ import {
   updateNumericFacetValues,
 } from '../../../../../features/facets/range-facets/numeric-facet-set/numeric-facet-actions';
 import {numericFacetSelectedValuesSelector} from '../../../../../features/facets/range-facets/numeric-facet-set/numeric-facet-selectors';
+import {numericFacetSetReducer as numericFacetSet} from '../../../../../features/facets/range-facets/numeric-facet-set/numeric-facet-set-slice';
+import {searchReducer as search} from '../../../../../features/search/search-slice';
 import {
   ConfigurationSection,
   FacetOptionsSection,
