+++ packages/headless/src/controllers/core/results-per-page/headless-core-results-per-page.ts
@@ -1,11 +1,11 @@
 import {NumberValue, Schema} from '@coveo/bueno';
 import {configuration} from '../../../app/common-reducers';
 import {CoreEngine} from '../../../app/engine';
-import {pagination} from '../../../app/reducers';
 import {
   registerNumberOfResults,
   updateNumberOfResults,
 } from '../../../features/pagination/pagination-actions';
+import {paginationReducer as pagination} from '../../../features/pagination/pagination-slice';
 import {
   ConfigurationSection,
   PaginationSection,
