+++ packages/headless/src/controllers/product-listing/category-facet/headless-product-listing-category-facet.ts
@@ -1,12 +1,9 @@
 import {configuration} from '../../../app/common-reducers';
 import {ProductListingEngine} from '../../../app/product-listing-engine/product-listing-engine';
-import {
-  categoryFacetSearchSet,
-  categoryFacetSet,
-  productListing,
-} from '../../../app/reducers';
+import {categoryFacetSetReducer as categoryFacetSet} from '../../../features/facets/category-facet-set/category-facet-set-slice';
 import {CategoryFacetSortCriterion} from '../../../features/facets/category-facet-set/interfaces/request';
 import {CategoryFacetValue} from '../../../features/facets/category-facet-set/interfaces/response';
+import {categoryFacetSearchSetReducer as categoryFacetSearchSet} from '../../../features/facets/facet-search-set/category/category-facet-search-set-slice';
 import {
   logFacetUpdateSort,
   logFacetShowMore,
@@ -16,6 +13,7 @@ import {
   logFacetSelect,
 } from '../../../features/facets/facet-set/facet-set-analytics-actions';
 import {fetchProductListing} from '../../../features/product-listing/product-listing-actions';
+import {productListingReducer as productListing} from '../../../features/product-listing/product-listing-slice';
 import {
   CategoryFacetSearchSection,
   CategoryFacetSection,
