+++ packages/headless/src/controllers/product-listing/range-facet/date-facet/headless-product-listing-date-filter.ts
@@ -1,11 +1,12 @@
 import {configuration} from '../../../../app/common-reducers';
 import {ProductListingEngine} from '../../../../app/product-listing-engine/product-listing-engine';
-import {dateFacetSet, search} from '../../../../app/reducers';
 import {
   logFacetClearAll,
   logFacetSelect,
 } from '../../../../features/facets/facet-set/facet-set-analytics-actions';
+import {dateFacetSetReducer as dateFacetSet} from '../../../../features/facets/range-facets/date-facet-set/date-facet-set-slice';
 import {fetchProductListing} from '../../../../features/product-listing/product-listing-actions';
+import {searchReducer as search} from '../../../../features/search/search-slice';
 import {
   ConfigurationSection,
   DateFacetSection,
