+++ packages/headless/src/controllers/product-listing/range-facet/numeric-facet/headless-product-listing-numeric-facet.ts
@@ -1,6 +1,5 @@
 import {configuration} from '../../../../app/common-reducers';
 import {ProductListingEngine} from '../../../../app/product-listing-engine/product-listing-engine';
-import {numericFacetSet, search} from '../../../../app/reducers';
 import {
   logFacetClearAll,
   logFacetUpdateSort,
@@ -9,7 +8,9 @@ import {RangeFacetSortCriterion} from '../../../../features/facets/range-facets/
 import {getAnalyticsActionForToggleRangeFacetSelect} from '../../../../features/facets/range-facets/generic/range-facet-utils';
 import {NumericRangeRequest} from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/request';
 import {NumericFacetValue} from '../../../../features/facets/range-facets/numeric-facet-set/interfaces/response';
+import {numericFacetSetReducer as numericFacetSet} from '../../../../features/facets/range-facets/numeric-facet-set/numeric-facet-set-slice';
 import {fetchProductListing} from '../../../../features/product-listing/product-listing-actions';
+import {searchReducer as search} from '../../../../features/search/search-slice';
 import {
   ConfigurationSection,
   NumericFacetSection,
