+++ packages/headless/src/controllers/product-listing/range-facet/numeric-facet/headless-product-listing-numeric-filter.ts
@@ -1,11 +1,12 @@
 import {configuration} from '../../../../app/common-reducers';
 import {ProductListingEngine} from '../../../../app/product-listing-engine/product-listing-engine';
-import {numericFacetSet, search} from '../../../../app/reducers';
 import {
   logFacetClearAll,
   logFacetSelect,
 } from '../../../../features/facets/facet-set/facet-set-analytics-actions';
+import {numericFacetSetReducer as numericFacetSet} from '../../../../features/facets/range-facets/numeric-facet-set/numeric-facet-set-slice';
 import {fetchProductListing} from '../../../../features/product-listing/product-listing-actions';
+import {searchReducer as search} from '../../../../features/search/search-slice';
 import {
   ConfigurationSection,
   NumericFacetSection,
